@@ -11,24 +11,27 @@
     "typecheck": "tsc"
   },
   "dependencies": {
+    "@plone/blocks": "workspace: *",
     "@plone/client": "workspace: *",
+    "@plone/components": "workspace: *",
     "@plone/providers": "workspace: *",
-    "@remix-run/css-bundle": "^2.4.0",
-    "@remix-run/node": "^2.8.1",
-    "@remix-run/react": "^2.8.1",
-    "@remix-run/serve": "^2.8.1",
-    "@tanstack/react-query": "^5.37.1",
+    "@plone/registry": "workspace: *",
+    "@remix-run/css-bundle": "^2.13.1",
+    "@remix-run/node": "^2.13.1",
+    "@remix-run/react": "^2.13.1",
+    "@remix-run/serve": "^2.13.1",
+    "@tanstack/react-query": "^5.59.0",
     "isbot": "^4.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "^2.8.1",
-    "@tanstack/react-query-devtools": "^5.37.1",
+    "@remix-run/dev": "^2.13.1",
+    "@tanstack/react-query-devtools": "^5.59.0",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
-    "typescript": "^5.4.2",
-    "vite": "^5.1.0",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.9",
     "vite-tsconfig-paths": "^4.2.1"
   },
   "engines": {
