@@ -36,14 +36,17 @@
     "husky": "9.0.11",
     "lint-staged": "15.2.2",
     "tsconfig": "workspace:*",
-    "typescript": "^5.4.5",
-    "vitest": "^1.5.0"
+    "typescript": "^5.6.3",
+    "vitest": "^2.1.3"
   },
   "packageManager": "pnpm@9.1.1",
   "pnpm": {
     "overrides": {
       "@pmmmwh/react-refresh-webpack-plugin": "0.5.11",
       "react-refresh": "0.14.0"
+    },
+    "patchedDependencies": {
+      "jest-resolve@26.6.2": "patches/jest-resolve@26.6.2.patch"
     }
   }
 }
