@@ -61,7 +61,7 @@
     "parcel": "^2.12.0",
     "release-it": "17.1.1",
     "tsconfig": "workspace:*",
-    "typescript": "5.4.2",
-    "vitest": "^1.3.1"
+    "typescript": "^5.6.3",
+    "vitest": "^2.1.3"
   }
 }
