@@ -8,7 +8,7 @@
     }
   ],
   "license": "MIT",
-  "version": "2.0.0-alpha.11",
+  "version": "2.0.0-alpha.16",
   "repository": {
     "type": "git",
     "url": "http://github.com/plone/components.git"
@@ -110,17 +110,17 @@
     "parcel": "^2.12.0",
     "release-it": "17.1.1",
     "storybook": "^8.0.4",
-    "typescript": "^5.4.5",
-    "vite": "^5.1.7",
-    "vitest": "^1.3.1",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.8",
+    "vitest": "^2.1.3",
     "vitest-axe": "^0.1.0"
   },
   "dependencies": {
-    "@react-aria/utils": "^3.22.0",
-    "@react-spectrum/utils": "^3.11.1",
+    "@react-aria/utils": "^3.25.3",
+    "@react-spectrum/utils": "^3.11.11",
     "@storybook/test": "^8.0.4",
-    "clsx": "^2.0.0",
-    "react-aria-components": "^1.2.0"
+    "clsx": "^2.1.1",
+    "react-aria-components": "^1.4.0"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
