@@ -9,7 +9,7 @@
   ],
   "funding": "https://github.com/sponsors/plone",
   "license": "MIT",
-  "version": "1.0.0-alpha.1",
+  "version": "1.0.0-alpha.3",
   "repository": {
     "type": "git",
     "url": "https://github.com/plone/volto.git"
@@ -43,7 +43,7 @@
   "scripts": {
     "watch": "parcel watch",
     "build": "parcel build",
-    "build:force": "parcel build --no-cache",
+    "build:force": "rm -rf dist && parcel build --no-cache",
     "test": "vitest",
     "dry-release": "release-it --dry-run",
     "release": "release-it",
@@ -62,7 +62,9 @@
   "dependencies": {
     "@plone/client": "workspace:*",
     "@plone/components": "workspace:*",
-    "@plone/registry": "workspace:*"
+    "@plone/registry": "workspace:*",
+    "@tanstack/react-query": "^5.59.0",
+    "react-aria-components": "^1.4.0"
   },
   "devDependencies": {
     "@parcel/config-default": "^2.12.0",
@@ -77,7 +79,7 @@
     "parcel": "^2.12.0",
     "release-it": "17.1.1",
     "tsconfig": "workspace:*",
-    "typescript": "5.2.2",
-    "vitest": "^1.3.1"
+    "typescript": "^5.4.5",
+    "vitest": "^2.1.3"
   }
 }
