@@ -9,7 +9,7 @@
   ],
   "funding": "https://github.com/sponsors/plone",
   "license": "MIT",
-  "version": "1.0.0-alpha.19",
+  "version": "1.0.0-alpha.21",
   "repository": {
     "type": "git",
     "url": "https://github.com/plone/volto.git"
@@ -56,6 +56,6 @@
     "react-intl": "3.12.1",
     "release-it": "^17.1.1",
     "tsconfig": "workspace:*",
-    "typescript": "5.4.2"
+    "typescript": "^5.6.3"
   }
 }
