@@ -1,11 +1,51 @@
-const fs = require('fs');
-const transform = require('@babel/core').transform;
-
-const getLoader = require('../../registry/src/create-addons-loader');
+import fs from 'fs';
+import { transform } from '@babel/core';
+import {
+  createAddonsLoader,
+  getAddonsLoaderCode,
+  nameFromPackage,
+} from '@plone/registry/create-addons-loader';
 
 describe('create-addons-loader code generation', () => {
-  test('no addon creates simple loader', () => {
-    const code = getLoader.getAddonsLoaderCode([]);
+  test('no addon creates simple loader, default = no loadProjectConfig', () => {
+    const code = getAddonsLoaderCode([]);
+    expect(code).toBe(`/*
+This file is autogenerated. Don't change it directly.
+Instead, change the "addons" setting in your package.json file.
+*/
+
+
+const addonsInfo = {};
+export { addonsInfo };
+
+const safeWrapper = (func) => (config) => {
+  const res = func(config);
+  if (typeof res === 'undefined') {
+    throw new Error("Configuration function doesn't return config");
+  }
+  return res;
+}
+
+const projectConfigLoader = false;
+const projectConfig = (config) => {
+  return projectConfigLoader && typeof projectConfigLoader.default === "function" ? projectConfigLoader.default(config) : config;
+}
+
+const load = (config) => {
+  const addonLoaders = [];
+  if(!addonLoaders.every((el) => typeof el === "function")) {
+    throw new TypeError(
+      'Each addon has to provide a function applying its configuration to the projects configuration.',
+    );
+  }
+  return projectConfig(addonLoaders.reduce((acc, apply) => safeWrapper(apply)(acc), config));
+};
+export default load;
+`);
+  });
+
+  test('no addon creates simple loader, loadProjectConfig set to true', () => {
+    const code = getAddonsLoaderCode([], {}, true);
     expect(code).toBe(`/*
 This file is autogenerated. Don't change it directly.
 Instead, change the "addons" setting in your package.json file.
@@ -24,8 +64,9 @@ const safeWrapper = (func) => (config) => {
   return res;
 }
 
+
 const projectConfig = (config) => {
-  return typeof projectConfigLoader.default === "function" ? projectConfigLoader.default(config) : config;
+  return projectConfigLoader && typeof projectConfigLoader.default === "function" ? projectConfigLoader.default(config) : config;
 }
 
 const load = (config) => {
@@ -42,17 +83,14 @@ export default load;
   });
 
   test('one addon creates loader', () => {
-    const code = getLoader.getAddonsLoaderCode(['volto-addon1']);
+    const code = getAddonsLoaderCode(['volto-addon1']);
     expect(code.indexOf("import voltoAddon1 from 'volto-addon1';") > 0).toBe(
       true,
     );
   });
 
   test('two addons create loaders', () => {
-    const code = getLoader.getAddonsLoaderCode([
-      'volto-addon1',
-      'volto-addon2',
-    ]);
+    const code = getAddonsLoaderCode(['volto-addon1', 'volto-addon2']);
     expect(
       code.indexOf(`
 import voltoAddon1 from 'volto-addon1';
@@ -61,7 +99,7 @@ import voltoAddon2 from 'volto-addon2';`) > 0,
   });
 
   test('one addons plus one extra creates loader', () => {
-    const code = getLoader.getAddonsLoaderCode(['volto-addon1:loadExtra1']);
+    const code = getAddonsLoaderCode(['volto-addon1:loadExtra1']);
     expect(
       code.indexOf(`
 import voltoAddon1, { loadExtra1 as loadExtra10 } from 'volto-addon1';
@@ -70,9 +108,7 @@ import voltoAddon1, { loadExtra1 as loadExtra10 } from 'volto-addon1';
   });
 
   test('one addons plus two extras creates loader', () => {
-    const code = getLoader.getAddonsLoaderCode([
-      'volto-addon1:loadExtra1,loadExtra2',
-    ]);
+    const code = getAddonsLoaderCode(['volto-addon1:loadExtra1,loadExtra2']);
     expect(
       code.indexOf(`
 import voltoAddon1, { loadExtra1 as loadExtra10, loadExtra2 as loadExtra21 } from 'volto-addon1';
@@ -81,7 +117,7 @@ import voltoAddon1, { loadExtra1 as loadExtra10, loadExtra2 as loadExtra21 } fro
   });
 
   test('two addons plus extras creates loader', () => {
-    const code = getLoader.getAddonsLoaderCode([
+    const code = getAddonsLoaderCode([
       'volto-addon1:loadExtra1,loadExtra2',
       'volto-addon2:loadExtra3,loadExtra4',
     ]);
@@ -95,7 +131,7 @@ import voltoAddon2, { loadExtra3 as loadExtra32, loadExtra4 as loadExtra43 } fro
 });
 
 describe('create-addons-loader default name generation', () => {
-  const getName = getLoader.nameFromPackage;
+  const getName = nameFromPackage;
 
   test('passing a simple word returns a word', () => {
     expect(getName('something')).toBe('something');
@@ -149,7 +185,7 @@ function makeAddonLoader(addons, load = true) {
       : require.resolve(name),
   );
 
-  const loaderPath = getLoader(addons);
+  const loaderPath = createAddonsLoader(addons);
   transpile(loaderPath);
 
   if (load) {
