@@ -1,37 +1,29 @@
 import { z } from 'zod';
 import { ApiRequestParams, apiRequest } from '../../API';
-import {
-  PloneClientConfig,
-  PloneClientConfigSchema,
-} from '../../validation/config';
+import { PloneClientConfig } from '../../validation/config';
 import { copyMoveDataSchema as moveDataSchema } from '../../validation/copymove';
 import { CopyMoveResponse as MoveResponse } from '@plone/types';
 
-export const MoveArgsSchema = z.object({
-  path: z.string(),
-  data: moveDataSchema,
-  config: PloneClientConfigSchema,
-});
-
-export type MoveArgs = z.infer<typeof MoveArgsSchema>;
+export type MoveArgs = z.infer<typeof moveDataSchema> & {
+  config: PloneClientConfig;
+};
 
 export const move = async ({
   path,
   data,
   config,
 }: MoveArgs): Promise<MoveResponse> => {
-  const validatedArgs = MoveArgsSchema.parse({
+  const validatedArgs = moveDataSchema.parse({
     path,
     data,
-    config,
   });
 
   const options: ApiRequestParams = {
+    config,
     data: validatedArgs.data,
-    config: validatedArgs.config,
   };
 
-  const movePath = `/${validatedArgs.path}/@move`;
+  const movePath = `${validatedArgs.path}/@move`;
 
   return apiRequest('post', movePath, options);
 };
