@@ -1,5 +1,8 @@
 import { z } from 'zod';
 
 export const copyMoveDataSchema = z.object({
-  source: z.union([z.string(), z.array(z.string())]),
+  path: z.string(),
+  data: z.object({
+    source: z.union([z.string(), z.array(z.string())]),
+  }),
 });
