@@ -1,10 +1,10 @@
 const path = require('path');
-const AddonConfigurationRegistry = require('../../registry/src/addon-registry');
+const { AddonRegistry } = require('@plone/registry/addon-registry');
 const WebpackRelativeResolver = require('../../volto/webpack-plugins/webpack-relative-resolver');
 
 const base = path.join(__dirname, '..');
 const makeRegistry = () => {
-  const registry = new AddonConfigurationRegistry(base);
+  const { registry } = AddonRegistry.init(base);
   registry.packages = {
     '@plone/volto-addon': {
       modulePath: '/somewhere/volto-addon/src',
