@@ -1,4 +1,4 @@
-describe('Add a new alias from control panel interface', () => {
+describe('URL Management control panel', () => {
   beforeEach(() => {
     cy.intercept('GET', `/**/*?expand*`).as('content');
     cy.autologin();
@@ -9,11 +9,28 @@ describe('Add a new alias from control panel interface', () => {
     });
     cy.visit('/my-page');
     cy.wait('@content');
+
+    cy.visit('/controlpanel/aliases');
+    cy.get('#add-alt-url').click();
+    cy.get('#field-altUrlPath').type('/alturl');
+    cy.get('#field-targetUrlPath').type('/my-page');
+    cy.get('button[aria-label="Save"]').click();
   });
   it('adds a new alias', () => {
+    cy.get('.toast-inner-content').contains('Alias has been added');
+  });
+  it('edits an existing new alias', () => {
     cy.visit('/controlpanel/aliases');
-    cy.get('#alternative-url-input').type('/alturl');
-    cy.get('#target-url-input').type('/my-page');
-    cy.get('#submit-alias').click();
+    cy.get('button[aria-label="Edit Alternative URL"]').click();
+    cy.get('#field-altUrlPath').clear().type('/alturl2');
+    cy.get('#field-targetUrlPath').clear().type('/my-page');
+    cy.get('button[aria-label="Save"]').click();
+    cy.get('.toast-inner-content').contains('Alias has been added');
+  });
+  it('removes an alias', () => {
+    cy.visit('/controlpanel/aliases');
+    cy.get('input[type="checkbox"][value="/alturl"]').click({ force: true });
+    cy.get('#remove-alt-urls').click();
+    cy.get('.toast-inner-content').contains('Aliases have been removed');
   });
 });
