@@ -80,9 +80,8 @@ const Login = (props) => {
   const loading = useSelector((state) => state.userSession.login.loading);
   const returnUrl =
     qs.parse(props.location?.search ?? location.search).return_url ||
-    location.pathname.replace(/\/login\/?$/, '').replace(/\/logout\/?$/, '') ||
+    location.pathname.replace(/\/[^/]*\/?$/, '') ||
     '/';
-
   const previousToken = usePrevious(token);
 
   useEffect(() => {
