@@ -18,11 +18,8 @@ export default function server() {
 
   server
     .listen(port, bind_address, () => {
-      if (app.apiPath === app.publicURL || !app.apiPath) {
-        console.log(`Volto is running in SEAMLESS mode`);
-      } else {
-        console.log(`API server (API_PATH) is set to: ${app.apiPath}`);
-      }
+      console.log(`API server (API_PATH) is set to: ${app.apiPath}`);
+
       if (app.devProxyToApiPath)
         console.log(
           `Proxying API requests from ${app.publicURL}/++api++ to ${
